open-ai-sdk-with-chainlit/pyproject.toml
[project]
name = "open-ai-sdk-with-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.16",
]
